moved hero to its own folder
